chore: realoc types #63
