+++ frontend/eventsync_front/src/stores/auth.ts
@@ -1,30 +1,8 @@
 import { defineStore } from 'pinia'
 import api from '@/services/api'
 import { jwtDecode, type JwtPayload } from 'jwt-decode'
+import {type User, type AuthState, type Profile} from '../types/users'
 
-export interface User {
-  email: string
-  password: string
-}
-
-interface AuthState {
-  user: User | null
-  token: string | null
-  rtoken: string | null
-  profile: Profile | null
-}
-
-interface Profile {
-  id: number
-  name: string
-  is_staff: boolean
-  birth_date: string
-  is_active: boolean
-  date_joined: string
-  cpf: string
-  phone: String
-  email: string
-}
 
 export const useAuthStore = defineStore({
   persist: true,
